Refactor NavLink, add boolean for course dropdown
