+++ src/AppBar/app-bar.less
@@ -65,7 +65,7 @@
       min-width: 176px;
       max-height: 0px;
       border: 1px solid @gray25;
-      border-top-color: transparent;
+      border-top: 0;
       box-shadow: 0 1px 2px @gray25;
 
       transform: translate(0, @nav-height);
@@ -192,11 +192,15 @@
     &.app-nav-link__toggle {
       height: @nav-height;
       position: absolute;
-      right: 30px;
+      right: 0px;
       top: 0;
       padding-right: 20px;
       padding-left: 20px;
       cursor: pointer;
+
+      @media @tablet {
+        right: 30px;
+      }
 
       @media @bigger {
         right: 0;
